requirements.txt
discord.py==1.3.4
dnspython==1.16.0